isolate dmi from the status file
