+++ src/daemon/status-file-product.h
@@ -1,0 +1,12 @@
+//
+// Created by costa on 4/4/25.
+//
+
+#ifndef NETDATA_STATUS_FILE_PRODUCT_H
+#define NETDATA_STATUS_FILE_PRODUCT_H
+
+#include "status-file.h"
+
+void product_name_vendor_type(DAEMON_STATUS_FILE *ds);
+
+#endif //NETDATA_STATUS_FILE_PRODUCT_H
